imakunai/jikken/requirements.txt
# requirements.txt
gradio
transformers
torch # または tensorflow, numpy など
paho-mqtt